Add transparency for CUENTA view at node cahabdd_s2
Add code to implement insert and delete DML operations to
CUENTA view for one node, this allows the user to do
these operations for CUENTA view without worrying about
Fragmentation Scheme, like if it were a table on a centralized database.

With this script, both nodes havae transparency
for CUENTA view.