Add transparency for MOVIMIENTO view at node cahabdd_s2
Add code to implement insert and delete DML operations to
MOVIMIENTO view for one node, this allows the user to do
these operations for MOVIMIENTO view without worrying about
Fragmentation Scheme, like if it were a table on a centralized database.

With this script, both nodes have transparency
for MOVIMIENTO view.